Refactor common code in webpack configuration
