+++ webpack/config.demo.js
@@ -2,16 +2,14 @@
 
 var path = require('path');
 var webpack = require('webpack');
-
-var ROOT = process.cwd();
-var SRC = path.join(ROOT, 'src');
-var DEMO = path.join(ROOT, 'demo');
+var moduleDefinition = require('./moduleDefinition');
+var directories = require('./directories');
 
 var NODE_ENV = process.env.NODE_ENV || 'development';
 var environment = JSON.stringify(NODE_ENV);
 
 var LIBRARY_NAME = 'dash_bootstrap_components';
-var BUILD_PATH = path.join(ROOT, 'demo-lib');
+var BUILD_PATH = path.join(directories.ROOT, 'demo-lib');
 
 var publicHost = process.env.DEMO_PUBLIC_HOST || undefined;
 
@@ -22,34 +20,9 @@ console.log('Using public host: ' + publicHost || '<default>');
 
 module.exports = {
   cache: false,
-  context: SRC,
+  context: directories.SRC,
   mode: NODE_ENV,
-  module: {
-    noParse: /node_modules\/json-schema\/lib\/validate\.js/, // used to get `request` to work: https://github.com/request/request/issues/1920#issuecomment-171246043
-    rules: [
-      {
-        test: /\.json$/,
-        loader: 'json-loader'
-      },
-      {
-        test: /\.jsx?$/,
-        include: [SRC, DEMO],
-        /*
-         * Use require.resolve to get a deterministic path
-         * and avoid webpack's magick loader resolution
-         */
-        use: [
-          {
-            loader: 'babel-loader'
-          }
-        ]
-      },
-      {
-        test: /\.css$/,
-        use: [{loader: 'style-loader'}, {loader: 'css-loader'}]
-      }
-    ]
-  },
+  module: moduleDefinition,
   plugins: [
     new webpack.DefinePlugin({
       'process.env': {
@@ -58,7 +31,7 @@ module.exports = {
     })
   ],
   entry: {
-    bundle: [path.join(ROOT, 'demo/index.js')]
+    bundle: [path.join(directories.ROOT, 'demo/index.js')]
   },
   devServer: {
     public: publicHost,
