+++ webpack/config.lib.js
@@ -2,18 +2,16 @@
 
 var path = require('path');
 var webpack = require('webpack');
+var moduleDefinition = require('./moduleDefinition');
+var directories = require('./directories')
 
 var OccurrenceOrderPlugin = require('webpack').optimize.OccurrenceOrderPlugin;
-
-var ROOT = process.cwd();
-var SRC = path.join(ROOT, 'src');
-var DEMO = path.join(ROOT, 'demo');
 
 var NODE_ENV = process.env.NODE_ENV || 'development';
 var environment = JSON.stringify(NODE_ENV);
 
 var LIBRARY_NAME = 'dash-bootstrap-components';
-var BUILD_PATH = path.join(ROOT, LIBRARY_NAME, '_components');
+var BUILD_PATH = path.join(directories.ROOT, LIBRARY_NAME, '_components');
 
 /* eslint-disable no-console */
 console.log('Current environment: ' + environment);
@@ -21,7 +19,7 @@ console.log('Current environment: ' + environment);
 
 module.exports = {
   cache: false,
-  context: SRC,
+  context: directories.SRC,
   mode: NODE_ENV,
   externals: [
     {
@@ -41,39 +39,7 @@ module.exports = {
       }
     }
   ],
-  module: {
-    noParse: /node_modules\/json-schema\/lib\/validate\.js/, // used to get `request` to work: https://github.com/request/request/issues/1920#issuecomment-171246043
-    rules: [
-      {
-        test: /\.json$/,
-        loader: 'json-loader'
-      },
-      {
-        test: /\.jsx?$/,
-        include: [SRC, DEMO],
-        /*
-         * Use require.resolve to get a deterministic path
-         * and avoid webpack's magick loader resolution
-         */
-        use: [
-          {
-            loader: 'babel-loader'
-          }
-        ]
-      },
-      {
-        test: /\.css$/,
-        use: [
-          {
-            loader: 'style-loader'
-          },
-          {
-            loader: 'css-loader'
-          }
-        ]
-      }
-    ]
-  },
+  module: moduleDefinition,
   plugins: [new OccurrenceOrderPlugin(true)],
   optimization: {
     minimize: true
@@ -84,7 +50,7 @@ module.exports = {
   output: {
     library: LIBRARY_NAME,
     libraryTarget: 'umd',
-    path: path.join(ROOT, 'lib/'),
+    path: path.join(directories.ROOT, 'lib/'),
     filename: LIBRARY_NAME + '.min.js',
     umdNamedDefine: true
   }
